Fixed two PostgreSQL bugs exposed by the strict GROUP BY test coverage
- Recently Viewed/Compared index collections compared visitor_id against an
  empty string when no visitor exists; PostgreSQL rejects '' for integer
  columns, and the failed statement aborted the surrounding test transaction.
  The condition values are now cast to int, so an absent visitor yields
  visitor_id = 0 (matches nothing), identical to MySQL's implicit coercion.
- The PostgreSQL reports helper built month/year rating-position periods with
  TO_CHAR, inserting a text expression into the date-typed period column.
  DATE_TRUNC('month'|'year', ...)::date keeps the expression date-typed.

+++ app/code/core/Mage/Reports/Model/Resource/Product/Index/Collection/Abstract.php
@@ -105,11 +105,13 @@ abstract class Mage_Reports_Model_Resource_Product_Index_Collection_Abstract ext
         $condition = [];
 
         if (Mage::getSingleton('customer/session')->isLoggedIn()) {
-            $condition['customer_id'] = Mage::getSingleton('customer/session')->getCustomerId();
+            $condition['customer_id'] = (int) Mage::getSingleton('customer/session')->getCustomerId();
         } elseif ($this->_customerId) {
-            $condition['customer_id'] = $this->_customerId;
+            $condition['customer_id'] = (int) $this->_customerId;
         } else {
-            $condition['visitor_id'] = Mage::getSingleton('log/visitor')->getId();
+            // Cast so an absent visitor yields visitor_id = 0 (matches nothing);
+            // an empty string is a type error on PostgreSQL integer columns.
+            $condition['visitor_id'] = (int) Mage::getSingleton('log/visitor')->getId();
         }
 
         return $condition;
